+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added Event-Driven (ED) jobs support for AMI installation
 - Added `STANDARD`, `SCHEDULED` to the `JobType` enum
+- Added description, deprecation date, and reason to the deprecated rules in the operational reports
 
 ### Fixed
 - Fixed invalid `next_token` handling